update bundle and install bcrypt
Bcrypt for has_secure_password
Forgot to add gemset and ruby version earlier. Adding
these causes some issues with the db and gems. Resolved
with this commit.

If issues with db not loading in Rails 4.0.2 console
can be resolved by doing a count on any model (User.count).
Seems that the db needs to be reset if gemsets are changed,
b/c the db was lost in mysql.
Reference: http://stackoverflow.com/questions/19755310/rails-
console-cant-connect-to-database/19774931#19774931